requirements.txt
# 依赖
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
